moved findOpenPosistion to an algorithms script
