+++ src/logic/updateState.js
@@ -33,13 +33,4 @@ const togglePlayer = (playerState) => {
   return "black";
 };
 
-const findOpenPosition = (state) => {
-  for (let counter = 1; counter < state.length; counter++) {
-    const element = state[counter];
-    if (element.team === null) {
-      return counter - 1;
-    }
-  }
-};
-
-module.exports = { resetState, startingState, togglePlayer, findOpenPosition };
+module.exports = { resetState, startingState, togglePlayer };
